python_ai_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
textstat
python-dotenv
asyncpg